feat: disables features conflicting with Istio
In order to handle creation of Notebooks which should belong to the service mesh we have to ensure this controller does not create unnecessary resources. For this purpose, when opendatahub.io/service-mesh annotation is present on the Notebook resource it will make sure that:

- no OAuth Proxy related resources are created
- no routes are created

It will fail admission of a notebook pod for the injection of proxy sidecars if annotation opendatahub.io/service-mesh is present and set to true as well.

I made few minor adjustements in the test code as I needed to re-use
several resources, such as NetworkPolicy.